Change data type of MemoryBytes to long

+++ app/src/main/java/uk/co/syski/client/android/data/entity/StorageEntity.java
@@ -17,5 +17,5 @@ public class StorageEntity {
 
     public String MemoryTypeName;
 
-    public int MemoryBytes;
+    public long MemoryBytes;
 }
